+++ src/components/Header/Header.module.css
@@ -1,0 +1,11 @@
+.navbar-bg {
+  background: #343a40;
+}
+
+.brand {
+  color: #ffffff;
+}
+
+.detail {
+  font-size: 15px;
+}
