+++ src/components/Header/index.js
@@ -1,0 +1,5 @@
+/**
+ * Re-export to prevent nested path
+ */
+import Header from './Header'
+export default Header
